Fixed bug in delete method by adding userid to Key and fixed return structure on the create method.

+++ serverless/project/completed-project/backend/src/lambda/http/deleteTodo.ts
@@ -2,6 +2,7 @@ import 'source-map-support/register'
 
 import { APIGatewayProxyHandler, APIGatewayProxyEvent, APIGatewayProxyResult } from 'aws-lambda'
 import * as AWS from 'aws-sdk'
+import { getUserId } from '../utils'
 
 const dynamoDocClient = new AWS.DynamoDB.DocumentClient()
 const TODOS_TABLE = process.env.TODOS_TABLE
@@ -10,9 +11,10 @@ const STATUS_OK = 200
 export const handler: APIGatewayProxyHandler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
 
     const todoID = event.pathParameters.todoId
+    const userID = getUserId(event)
 
     //https://docs.aws.amazon.com/AWSJavaScriptSDK/latest/AWS/DynamoDB/DocumentClient.html#delete-property
-    await dynamoDocClient.delete({ TableName: TODOS_TABLE, Key: { todoId: todoID } }).promise()
+    await dynamoDocClient.delete({ TableName: TODOS_TABLE, Key: { todoId: todoID , userId: userID} }).promise()
 
     // https://stackoverflow.com/questions/2342579/http-status-code-for-update-and-delete
     return { statusCode: STATUS_OK,
